==> Code/1 - Model Selection.Rmd <==
---
title: Abiotic factors modify ponderosa pine regeneration outcomes after high-severity fire - Hurdle model selection 
author: "Kevin Willson"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    toc: true
    number_sections: true
    toc_depth: 5
    code_folding: show
    #df_print: paged
    #df_print: kable
    #toc_float: true
      #collapsed: false
      #smooth_scroll: TRUE
    theme: cosmo #spacelab #yeti #united
    highlight: tango
  pdf_document:
    df_print: kable
    #latex_engine: xelatex
#sansfont: IBM Plex Sans
#classoption: landscape
fontsize: 12pt
geometry: margin=0.25in
always_allow_html: yes
---

# This HTML outlines the model selection process to determine the best model for estimating regeneration density using topographic and wind predictor variables.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, include=FALSE, cache = FALSE, message=FALSE, warning=FALSE}
library(readxl)
library(dplyr)
library(bimixt)
library(car)
library(MASS)
library(tidyr)
library(data.table)
library(forecast)
library(lmtest)
library(MuMIn)
library(caret)
library(broom)
library(rstatix)
library(gridExtra)
library(DescTools)
library(knitr)
library(ggplot2)
library(ggpubr)
library(statmod)
library(DHARMa)
```


# Prepare data for analysis
```{r}
path <- "D:/Dryad_submission/Data/"
seedling_distance <- read_excel(paste0(path,"Transect_Data_5-25-21.xlsx"), sheet = 1)
Plotdata_HA <- read_excel(paste0(path,"Plot_data.xlsx"), sheet = 1)
seedling_distance$FPT <- paste(seedling_distance$Fire, "-", seedling_distance$Patch, "-", seedling_distance$Transect)
seedling_distance <- seedling_distance[!is.na(seedling_distance$Vertical_Distance),]

## Remove transects that had nearby overstory trees at the far end of the transect
seedling_distance <- seedling_distance[!(seedling_distance$Fire == "Shelly" & seedling_distance$Patch == "NB" & seedling_distance$Transect == "2"),]
seedling_distance <- seedling_distance[!(seedling_distance$Fire == "Shelly" & seedling_distance$Patch == "1" & seedling_distance$Transect == "2"),]
seedling_distance <- seedling_distance[!(seedling_distance$Fire == "Glass" & seedling_distance$Patch == "2" & seedling_distance$Transect == "7"),]
seedling_distance <- seedling_distance[!(seedling_distance$Fire == "Shelly" & seedling_distance$Patch == "NB" & seedling_distance$Transect == "1"),]
seedling_distance <- seedling_distance[!(seedling_distance$Fire == "Divide" & seedling_distance$Patch == "6" & seedling_distance$Transect == "10"),]

Plotdata_HA <- Plotdata_HA[!(Plotdata_HA$Fire == "Shelly" & Plotdata_HA$Patch == "NB" & Plotdata_HA$Transect == "2"),]
Plotdata_HA <- Plotdata_HA[!(Plotdata_HA$Fire == "Shelly" & Plotdata_HA$Patch == "1" & Plotdata_HA$Transect == "2"),]
Plotdata_HA <- Plotdata_HA[!(Plotdata_HA$Fire == "Glass" & Plotdata_HA$Patch == "2" & Plotdata_HA$Transect == "7"),]
Plotdata_HA <- Plotdata_HA[!(Plotdata_HA$Fire == "Shelly" & Plotdata_HA$Patch == "NB" & Plotdata_HA$Transect == "1"),]
Plotdata_HA <- Plotdata_HA[!(Plotdata_HA$Fire == "Divide" & Plotdata_HA$Patch == "6" & Plotdata_HA$Transect == "10"),]

## Get equal densities across all transects (just shelly 1-1 that was run at 1 meter width instead of 10 meters)
x <- 1
additional_shelly_1_1 <- seedling_distance[FALSE,]
d <- seedling_distance[FALSE,]

while(x <= length(seedling_distance$Fire)){
  if(seedling_distance$Fire[x] == "Shelly" & seedling_distance$Patch[x] == 1 & seedling_distance$Transect[x] == 1){
    d <- rbind(seedling_distance[x,],seedling_distance[x,],seedling_distance[x,],seedling_distance[x,],seedling_distance[x,],
               seedling_distance[x,],seedling_distance[x,],seedling_distance[x,],seedling_distance[x,])
    additional_shelly_1_1 <- rbind(additional_shelly_1_1, d)
    x <- x + 1
  }
  else{
    x <- x + 1
  }
}

seedling_distance <- rbind(seedling_distance, additional_shelly_1_1)
```

## Convert data to densities
```{r}
### Seedling density is quantified by counting the number of seedlings in 16 10x10 meter blocks starting from 0-10 meters from the intact forest and ending at 150-160 meters for each transect. Those values were then converted to seedling density at the hectare scale by multiplying by 100 (100 sq m to 10,000 sq m = 1 ha).

x <- 1
seedling_distance$Distance_grouping <- NA
while(x <= length(seedling_distance$Fire)){
  if(seedling_distance$Vertical_Distance[x] <= 10){
    seedling_distance$Distance_grouping[x] <- 5
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 10 & seedling_distance$Vertical_Distance[x] <= 20){
    seedling_distance$Distance_grouping[x] <- 15
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 20 & seedling_distance$Vertical_Distance[x] <= 30){
    seedling_distance$Distance_grouping[x] <- 25
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 30 & seedling_distance$Vertical_Distance[x] <= 40){
    seedling_distance$Distance_grouping[x] <- 35
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 40 & seedling_distance$Vertical_Distance[x] <= 50){
    seedling_distance$Distance_grouping[x] <- 45
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 50 & seedling_distance$Vertical_Distance[x] <= 60){
    seedling_distance$Distance_grouping[x] <- 55
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 60 & seedling_distance$Vertical_Distance[x] <= 70){
    seedling_distance$Distance_grouping[x] <- 65
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 70 & seedling_distance$Vertical_Distance[x] <= 80){
    seedling_distance$Distance_grouping[x] <- 75
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 80 & seedling_distance$Vertical_Distance[x] <= 90){
    seedling_distance$Distance_grouping[x] <- 85
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 90 & seedling_distance$Vertical_Distance[x] <= 100){
    seedling_distance$Distance_grouping[x] <- 95
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 100 & seedling_distance$Vertical_Distance[x] <= 110){
    seedling_distance$Distance_grouping[x] <- 105
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 110 & seedling_distance$Vertical_Distance[x] <= 120){
    seedling_distance$Distance_grouping[x] <- 115
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 120 & seedling_distance$Vertical_Distance[x] <= 130){
    seedling_distance$Distance_grouping[x] <- 125
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 130 & seedling_distance$Vertical_Distance[x] <= 140){
    seedling_distance$Distance_grouping[x] <- 135
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 140 & seedling_distance$Vertical_Distance[x] <= 150){
    seedling_distance$Distance_grouping[x] <- 145
    x <- x+1
  }
  else if(seedling_distance$Vertical_Distance[x] > 150 & seedling_distance$Vertical_Distance[x] <= 160){
    seedling_distance$Distance_grouping[x] <- 155
    x <- x+1
  }
}
```

## Group these individuals by transect and distance to count the number of each
```{r}
seed_density_distance <- seedling_distance %>% 
  group_by(Fire, Patch, Transect, Distance_grouping) %>% 
  tally()

unique_transects <- Plotdata_HA[,c(1:3)]

all_bins <- unique_transects[rep(seq_len(nrow(unique_transects)), each = 16), ]

all_bins$Distance_grouping <- rep(seq(from = 5, to = 155, by = 10), 46)

all_bins$n_zeros <- 0

seed_density_distance <- merge(seed_density_distance, all_bins, all.y = T)

seed_density_distance[is.na(seed_density_distance)] <- 0

seed_density_distance$n_zeros <- NULL

## Convert density to ha (100 sq m to 10,000 sq m)
seed_density_distance$Seedling_density <- seed_density_distance$n*100

## Bind the plot data with it in preparation for analysis
merging <- Plotdata_HA[, c(1,2,3, 8, 9, 10, 11, 6, 12, 13, 4, 5, 7)]

seed_density_distance <- merge(seed_density_distance, merging, by = c("Fire", "Patch", "Transect"), all.x = T)

seed_density_distance <- seed_density_distance[!is.na(seed_density_distance$AvgHeight),]

colnames(seed_density_distance)[4] <- "Distance"
colnames(seed_density_distance)[6] <- "Seedling_density"
colnames(seed_density_distance)[8] <- "Combined_slope"
colnames(seed_density_distance)[9] <- "Aspect"
colnames(seed_density_distance)[10] <- "Slope_Angle"
colnames(seed_density_distance)[11] <- "Tree_height"
colnames(seed_density_distance)[12] <- "Prevailing_Wind"
colnames(seed_density_distance)[13] <- "Cont_Prevailing_Wind"
colnames(seed_density_distance)[15] <- "Ponderosa_overstory_density" 

### For some reason, the prevailing wind value in the dataframe was incorrect, this corrects it
seed_density_distance$Prevailing_Wind <- ifelse(seed_density_distance$Cont_Prevailing_Wind>0, "Yes", "No")
seed_density_distance$Cont_Prevailing_Wind_0 <- ifelse(seed_density_distance$Cont_Prevailing_Wind>0, seed_density_distance$Cont_Prevailing_Wind,0)
```

# Wind variable selection 
```{r}
slopecomb_wind <- glm(Seedling_density ~  Cont_Prevailing_Wind + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )

slopecomb_wind_0 <- glm(Seedling_density ~ Cont_Prevailing_Wind_0 + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
```

## View table
``` {r}
msAICc_wind <- model.sel(slopecomb_wind, slopecomb_wind_0)
kable(msAICc_wind)
```

### Cutting off wind at 0 worked better
``` {r}
seed_density_distance$Cont_Prevailing_Wind <- seed_density_distance$Cont_Prevailing_Wind_0

seed_density_distance$Cont_Prevailing_Wind_0 <- NULL

### Save this dataframe for making figures (script 4a - Figures Script.R)
write.csv(seed_density_distance, "final_model_df.csv")

```

# Specify binomial models and determine which is the most parsimonious
## The binomial and general linear model can be different in creating the hurdle model
#### https://stats.stackexchange.com/questions/459207/two-part-model-is-it-necessary-to-use-the-same-regressors-in-both-parts
#### https://www.stata.com/stata14/hurdle-models/
#### https://seananderson.ca/2014/05/18/gamma-hurdle/
#### Andika, A., Abdullah, S., & Nurrohmah, S. (2019). Hurdle Negative Binomial Regression Model. In ICSA-International Conference on Statistics and Analytics 2019 (pp. 57-68). 
```{r}
seed_density_distance$binom <- ifelse(seed_density_distance$Seedling_density > 0,1,0)

comb_slope_binom <- glm(binom ~ Distance + factor(Aspect) + Combined_slope, family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_treeht_binom <- glm(binom ~ Combined_slope + Tree_height + Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_wind_binom <- glm(binom ~  Combined_slope + Cont_Prevailing_Wind + Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_treeht_wind_binom <- glm(binom ~ Combined_slope + Tree_height + Cont_Prevailing_Wind + Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

treeht_wind_binom <- glm(binom ~  Tree_height + Cont_Prevailing_Wind+ Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

treeht_binom <- glm(binom ~ Tree_height+ Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

wind_binom <- glm(binom ~ Cont_Prevailing_Wind+ Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_pipodensity25_binom <- glm(binom ~  Combined_slope + Ponderosa_overstory_density+ Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_wind_pipodensity25_binom <- glm(binom ~  Combined_slope + Ponderosa_overstory_density + Cont_Prevailing_Wind+ Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

pipodensity25_binom <- glm(binom ~ Ponderosa_overstory_density + Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

pipodensity25_wind_binom <- glm(binom ~ Ponderosa_overstory_density + Cont_Prevailing_Wind + Distance + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

```

## View table

``` {r}
msAICc_binom <- model.sel(comb_slope_binom, treeht_binom, wind_binom, pipodensity25_binom, comb_slope_treeht_binom, comb_slope_wind_binom, comb_slope_pipodensity25_binom, treeht_wind_binom, pipodensity25_wind_binom, comb_slope_treeht_wind_binom, comb_slope_wind_pipodensity25_binom)

## The + signs only indicate that the factor variable was included in the model.
kable(msAICc_binom)
```

## Run diagnostics on the best binomial model
### Followed directions from here:
#### http://www.sthda.com/english/articles/36-classification-methods-essentials/148-logistic-regression-assumptions-and-diagnostics-in-r/
```{r}
probabilities <- predict(comb_slope_wind_pipodensity25_binom, type = "response")
predicted.classes <- ifelse(probabilities > 0.5, "pos", "neg")
```

### Linearity assumption
```{r}
# Select only numeric predictors
mydata <- seed_density_distance %>%
  select(Combined_slope, Ponderosa_overstory_density, Cont_Prevailing_Wind, Distance) 

predictors <- colnames(mydata)

# Bind the logit and tidying the data for plot
mydata <- mydata %>%
  mutate(logit = log(probabilities/(1-probabilities))) %>%
  gather(key = "predictors", value = "predictor.value", -logit)

ggplot(mydata, aes(logit, predictor.value))+
  geom_point(size = 0.5, alpha = 0.5) +
  geom_smooth(method = "loess") + 
  theme_bw() + 
  facet_wrap(~predictors, scales = "free_y")
```

### wind and bin distance seems to be non-linear, might need to do a transformation on one of them
```{r}
comb_slope_wind_pipodensity25_binom <- glm(binom ~  Combined_slope + Ponderosa_overstory_density + (Cont_Prevailing_Wind) + log(Distance) + factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )
```

### Try it again
```{r}
probabilities <- predict(comb_slope_wind_pipodensity25_binom, type = "response")
predicted.classes <- ifelse(probabilities > 0.5, "pos", "neg")
```

### Linearity assumption
```{r}
# Select only numeric predictors
mydata <- seed_density_distance %>%
  select(Combined_slope, Ponderosa_overstory_density, Cont_Prevailing_Wind, Distance) 

predictors <- colnames(mydata)

# Bind the logit and tidying the data for plot
mydata <- mydata %>%
  mutate(logit = log(probabilities/(1-probabilities))) %>%
  gather(key = "predictors", value = "predictor.value", -logit)

ggplot(mydata, aes(logit, predictor.value))+
  geom_point(size = 0.5, alpha = 0.5) +
  geom_smooth(method = "loess") + 
  theme_bw() + 
  facet_wrap(~predictors, scales = "free_y")
```

#### This is better, lets add this to all the models 

```{r}
comb_slope_binom <- glm(binom ~log(Distance)+ factor(Aspect) + Combined_slope, family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_treeht_binom <- glm(binom ~ Combined_slope + Tree_height +log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_wind_binom <- glm(binom ~  Combined_slope + Cont_Prevailing_Wind +log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_treeht_wind_binom <- glm(binom ~ Combined_slope + Tree_height + Cont_Prevailing_Wind +log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

treeht_wind_binom <- glm(binom ~  Tree_height + Cont_Prevailing_Wind+log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

treeht_binom <- glm(binom ~ Tree_height+log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

wind_binom <- glm(binom ~ Cont_Prevailing_Wind+log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_pipodensity25_binom <- glm(binom ~  Combined_slope + Ponderosa_overstory_density+log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

comb_slope_wind_pipodensity25_binom <- glm(binom ~  Combined_slope + Ponderosa_overstory_density + Cont_Prevailing_Wind+log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

pipodensity25_binom <- glm(binom ~ Ponderosa_overstory_density +log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

pipodensity25_wind_binom <- glm(binom ~ Ponderosa_overstory_density + Cont_Prevailing_Wind +log(Distance)+ factor(Aspect), family = binomial(), data=seed_density_distance,  na.action = "na.fail" )

```

## View table

``` {r}
msAICc_binom <- model.sel(comb_slope_binom, treeht_binom, wind_binom, pipodensity25_binom, comb_slope_treeht_binom, comb_slope_wind_binom, comb_slope_pipodensity25_binom, treeht_wind_binom, pipodensity25_wind_binom, comb_slope_treeht_wind_binom, comb_slope_wind_pipodensity25_binom)

## The + signs only indicate that the factor variable was included in the model.
kable(msAICc_binom)
```


### Influential values 
```{r}
plot(comb_slope_wind_pipodensity25_binom, which = 4, id.n = 3)
```

#### None are above 0.5, none of these values are influential outliers

### Multicollinearity
```{r}
vif(comb_slope_wind_pipodensity25_binom)
```

#### All values are below 4, no issues with correlation between predictor variables

## Check the regression stats
```{r}
summary(comb_slope_wind_pipodensity25_binom)
PseudoR2(comb_slope_wind_pipodensity25_binom, which = "CoxSnell")
```

### a small check on outputs from two specific models for extrapolation
#### wind-only
```{r}
summary(wind_binom)
```

#### wind and slope
```{r}
summary(comb_slope_wind_binom)
```

# Specify general linear models with the nonzero data and determine which is the most parsimonious
```{r}
seed_density_distance <- seed_density_distance[seed_density_distance$binom == 1,]

comb_slope <- glm(Seedling_density ~ Distance + factor(Aspect) + Combined_slope, data=seed_density_distance,  na.action = "na.fail" )
comb_slope_treeht <- glm(Seedling_density ~ Combined_slope + Tree_height + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_wind<- glm(Seedling_density ~  Combined_slope + Cont_Prevailing_Wind + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_treeht_wind<- glm(Seedling_density ~ Combined_slope + Tree_height + Cont_Prevailing_Wind + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
treeht_wind<- glm(Seedling_density ~  Tree_height + Cont_Prevailing_Wind+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
treeht<- glm(Seedling_density ~ Tree_height+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
wind<- glm(Seedling_density ~ Cont_Prevailing_Wind+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_pipodensity25 <- glm(Seedling_density ~  Combined_slope + Ponderosa_overstory_density+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_wind_pipodensity25 <- glm(Seedling_density ~  Combined_slope + Ponderosa_overstory_density + Cont_Prevailing_Wind+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
pipodensity25 <- glm(Seedling_density ~ Ponderosa_overstory_density + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
pipodensity25_wind <- glm(Seedling_density ~ Ponderosa_overstory_density + Cont_Prevailing_Wind + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
```

## View table

``` {r}
msAICc <- model.sel(comb_slope, treeht, wind, pipodensity25, comb_slope_treeht, comb_slope_wind, comb_slope_pipodensity25, treeht_wind, pipodensity25_wind, comb_slope_treeht_wind, comb_slope_wind_pipodensity25)

## The + signs only indicate that the factor variable was included in the model.
kable(msAICc)
```

### Check 

## Run diagnostics on the best model
### Check qq plot and shapiro test

```{r}
qqnorm(qresid(comb_slope_wind_pipodensity25)); qqline(qresid(comb_slope_wind_pipodensity25))

shapiro.test(comb_slope_wind_pipodensity25$residuals)

#### Not normal
```

### Check the residual distribution 

```{r}
plot(density(rstandard(comb_slope_wind_pipodensity25, type='pearson')))

## This model did not meet normal residuals assumption, so need to transform the data
```



## Log transformation. Close to normal
```{r}
log_Density <- glm(log(Seedling_density) ~ Combined_slope + Ponderosa_overstory_density + Cont_Prevailing_Wind + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
qqnorm(qresid(log_Density)); qqline(qresid(log_Density))
plot(density(rstandard(log_Density, type='pearson')))

shapiro.test(log_Density$residuals)
```

### Lets check to see if any predictor transformations improve normality
```{r}
#### I did a polynomial transformation on Overstory pipo density and prevailing wind to get normality

log_Density <- glm(log(Seedling_density) ~ Combined_slope + poly(Ponderosa_overstory_density,2) + poly(Cont_Prevailing_Wind,2) + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
qqnorm(qresid(log_Density)); qqline(qresid(log_Density))
plot(density(rstandard(log_Density, type='pearson')))

shapiro.test(log_Density$residuals)

### Normality!
```


# Apply transformations to all models and re-run the analysis
```{r}
comb_slope <- glm(log(Seedling_density) ~ Distance + factor(Aspect) + Combined_slope, data=seed_density_distance,  na.action = "na.fail" )
comb_slope_treeht <- glm(log(Seedling_density) ~ Combined_slope + Tree_height + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_wind<- glm(log(Seedling_density) ~  Combined_slope + poly(Cont_Prevailing_Wind,2)+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_treeht_wind<- glm(log(Seedling_density) ~ Combined_slope + Tree_height + poly(Cont_Prevailing_Wind,2) + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
treeht_wind<- glm(log(Seedling_density) ~  Tree_height + poly(Cont_Prevailing_Wind,2)+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
treeht<- glm(log(Seedling_density) ~ Tree_height+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
wind<- glm(log(Seedling_density) ~ poly(Cont_Prevailing_Wind,2)+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_pipodensity25 <- glm(log(Seedling_density) ~  Combined_slope + poly(Ponderosa_overstory_density,2)+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
comb_slope_wind_pipodensity25 <- glm(log(Seedling_density) ~  Combined_slope + poly(Ponderosa_overstory_density,2) + poly(Cont_Prevailing_Wind,2)+ Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
pipodensity25 <- glm(log(Seedling_density) ~ poly(Ponderosa_overstory_density,2) + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
pipodensity25_wind <- glm(log(Seedling_density) ~ poly(Ponderosa_overstory_density,2) + poly(Cont_Prevailing_Wind,2) + Distance + factor(Aspect), data=seed_density_distance,  na.action = "na.fail" )
```

## View table
``` {r}
msAICc_log <- model.sel(comb_slope, treeht, wind, pipodensity25, comb_slope_treeht, comb_slope_wind, comb_slope_pipodensity25, treeht_wind, pipodensity25_wind, comb_slope_treeht_wind, comb_slope_wind_pipodensity25)
kable(msAICc_log)

## After transformation, only one model has a delta AIC <= 2.
```



### Normality check for Slope, overstory PIPO density, wind, aspect, distance to intact forest model
```{r}
qqnorm(qresid(comb_slope_wind_pipodensity25)); qqline(qresid(comb_slope_wind_pipodensity25))

plot(density(rstandard(comb_slope_wind_pipodensity25, type='pearson')))

shapiro.test(comb_slope_wind_pipodensity25$residuals)

### Passes the normality check. Lets check the other assumptions for this model 
```

### Check for homogeneity of variance, no flagrant trends
```{r}
scatter.smooth(1:288, rstandard(comb_slope_wind_pipodensity25, type='deviance'), col='gray') 
```

### Fitted vs. Residuals graph 
```{r}
scatter.smooth(predict(comb_slope_wind_pipodensity25, type='response'), qresid(comb_slope_wind_pipodensity25), col='gray') 
```

### Residuals vs. predictor variables graph
```{r}
plot(seed_density_distance$Ponderosa_overstory_density, resid(comb_slope_wind_pipodensity25))
abline(0,0)

plot((seed_density_distance$Ponderosa_overstory_density)^2, resid(comb_slope_wind_pipodensity25))
abline(0,0)

plot(seed_density_distance$Cont_Prevailing_Wind, resid(comb_slope_wind_pipodensity25))
abline(0,0)

plot((seed_density_distance$Cont_Prevailing_Wind)^2, resid(comb_slope_wind_pipodensity25))
abline(0,0)

plot(seed_density_distance$Distance, resid(comb_slope_wind_pipodensity25))
abline(0,0)
```

### Check for multicollinearity
```{r}
vif(comb_slope_wind_pipodensity25) 
## No collinearity 
```

### Check link function
```{r}
plot(predict(comb_slope_wind_pipodensity25), resid(comb_slope_wind_pipodensity25, type = "working"), col='gray') #
abline(0,0)
```

### Check for spatial autocorrelation in the residuals using a Moran's I test
```{r}
seed_density_distance$FPT <- paste0(seed_density_distance$Fire, seed_density_distance$Patch, seed_density_distance$Transect)

fpt_unique <- unique(seed_density_distance$FPT)

fpt_unique <- cbind(fpt_unique, 1:length(fpt_unique))

seed_density_distance <- merge(seed_density_distance, fpt_unique, by.x = "FPT", by.y = "fpt_unique", all.x = T)

res = simulateResiduals(comb_slope_wind_pipodensity25)

testSpatialAutocorrelation(res, x =  seed_density_distance$V2, y = seed_density_distance$Distance)

## Not significant, indicating no spatial autocorrelation within the residuals among all transects

seed_density_distance$FPT_num <- as.numeric(as.factor(seed_density_distance$FPT))


res2 = recalculateResiduals(res, group = seed_density_distance$FPT_num)

testSpatialAutocorrelation(res2, 
                           x =  aggregate(as.numeric(seed_density_distance$V2), list(seed_density_distance$FPT_num), mean)$x, 
                           y = aggregate(seed_density_distance$Distance, list(seed_density_distance$FPT_num), mean)$x)


## Not significant, indicating no spatial autocorrelation among aggregated transect residuals

seed_density_distance$Fire_patch <- as.numeric(as.factor(paste0(seed_density_distance$Fire, seed_density_distance$Patch)))


res3 = recalculateResiduals(res, group = seed_density_distance$Fire_patch)

testSpatialAutocorrelation(res3, 
                           x =  aggregate(as.numeric(seed_density_distance$V2), list(seed_density_distance$Fire_patch), mean)$x, 
                           y = aggregate(seed_density_distance$Distance, list(seed_density_distance$Fire_patch), mean)$x)

## Not significant, indicating no spatial autocorrelation among all transects by patch


```

### Check linearity 
```{r}
ggplot(seed_density_distance,aes(y=log(Seedling_density), x=comb_slope_wind_pipodensity25$fitted.values)) +
  geom_point() +
  geom_smooth(method = "lm")+
  stat_regline_equation(label.y.npc = .125, label.x.npc = .6) +
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~','~")), label.y.npc = 0.00, label.x.npc = .6)+
  theme(panel.border = element_blank(), panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(color = "black"), axis.title = element_text(size =25),
        axis.text = element_text(size = 15))+
  xlab("Predicted Seed Density (log(seeds/ha))")+ ylab("Actual  Seed Density (log(seeds/ha))") +
  ggtitle( expression( "Predicted versus actual Seed Density " )) + theme(plot.title = element_text(hjust = 0.5, size = 25))
```

### Check the individual variable relationships
```{r}
avPlots(comb_slope_wind_pipodensity25)
```

## Check the regression stats
```{r}
summary(comb_slope_wind_pipodensity25)
PseudoR2(comb_slope_wind_pipodensity25, which = "CoxSnell")
```

### a small check on outputs from the model for extrapolation
#### wind and slope
```{r}
summary(comb_slope_wind)
```
